fix test for MediaFile
